- Purposely broke the association between Postfix's DB style and Dystill, hopefully allowing for wider adoption.
  - This means dropping the user_id association in Dystill's DB schema
and replacing it with a text field containing the user email
address. "Wildcard" rules are still supported using an empty string
('') in the email field.
    
  - New configuration parameter, maildir_path, specifies the location
and path of the maildir. {to_address} is replaced with the email
address of the recipient (the maildir location)
    
- BUGFIX: An issue where folder names with certain characters would
cause issues with mail delivery.Signed-off-by: Rebecca Peck
<me@rebeccapeck.org>